Issue #11221: Prepping for batchmode tests
